% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bootstrap_auc.R
\name{plot_bootstrap_auc}
\alias{plot_bootstrap_auc}
\title{Plot Bootstrap AUC Distribution}
\usage{
plot_bootstrap_auc(bootstrap_result, bins = 30)
}
\arguments{
\item{bootstrap_result}{Object returned from \code{bootstrap_auc()}.}

\item{bins}{Number of histogram bins.}
}
\value{
A ggplot object.
}
\description{
Creates a visualization of the bootstrap distribution of AUC estimates.
The plot displays the uncertainty distribution, estimated AUC,
and bootstrap confidence interval boundaries.
}
\examples{
fit <- glm(
  am ~ mpg + hp,
  data = mtcars,
  family = binomial
)

result <- bootstrap_auc(
  fit,
  mtcars,
  "am",
  B = 100
)

plot_bootstrap_auc(result)

}
